requirements.txt

streamlit==1.28.1
PyMuPDF==1.23.8
python-dotenv==1.0.0
reportlab==4.0.4
requests==2.31.0
transformers==4.35.2
sentencepiece==0.1.99
protobuf==3.20.3
torch>=2.0.0
EOL